requirements.txt
anthropic
fire
gcsfs
google-aiplatform
google-generativeai
google-cloud-aiplatform
google-cloud-storage
huggingface-hub==0.30.1
openai
sentencepiece
tiktoken
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
transformers==4.50.3
vertexai
vllm==0.8.2